DRAW-110 fix: Api - Exception 모듈 연동
